refactor(navigation): update files structure
